Reorganize simple path finder
This commit extracts the simple path finder into its own package,
simple path, along with other source reorganization and refactoring
