+++ src/github.com/stellar/horizon/paths/main.go
@@ -4,6 +4,7 @@ import (
 	"github.com/stellar/go-stellar-base/xdr"
 )
 
+// Query is a query for paths
 type Query struct {
 	DestinationAddress string
 	DestinationAsset   xdr.Asset
@@ -11,53 +12,18 @@ type Query struct {
 	SourceAssets       []xdr.Asset
 }
 
+// Path is the interface that represents a single result returned
+// by a path finder.
 type Path interface {
 	Path() []xdr.Asset
 	Source() xdr.Asset
 	Destination() xdr.Asset
-
 	// Cost returns an amount (which may be estimated), delimited in the Source assets
 	// that is suitable for use as the `sendMax` field for a `PathPaymentOp` struct.
 	Cost(amount xdr.Int64) xdr.Int64
 }
 
+// Finder finds paths.
 type Finder interface {
 	Find(Query) ([]Path, error)
 }
-
-type DummyFinder struct {
-}
-
-func (f *DummyFinder) Find(q Query) ([]Path, error) {
-	paths := make([]Path, 2)
-	n, err := xdr.NewAsset(xdr.AssetTypeAssetTypeNative, nil)
-
-	if err != nil {
-		return nil, err
-	}
-
-	paths[0] = DummyPath{
-		source:      n,
-		destination: n,
-		path:        []xdr.Asset{n, n, n},
-	}
-
-	paths[1] = DummyPath{
-		source:      n,
-		destination: n,
-		path:        []xdr.Asset{n, n, n},
-	}
-
-	return paths, nil
-}
-
-type DummyPath struct {
-	source      xdr.Asset
-	destination xdr.Asset
-	path        []xdr.Asset
-}
-
-func (d DummyPath) Source() xdr.Asset               { return d.source }
-func (d DummyPath) Destination() xdr.Asset          { return d.destination }
-func (d DummyPath) Path() []xdr.Asset               { return d.path }
-func (d DummyPath) Cost(amount xdr.Int64) xdr.Int64 { return amount }
