+++ src/github.com/stellar/horizon/simplepath/doc.go
@@ -1,0 +1,3 @@
+// Package simplepath provides an implementation of paths.Finder that performs
+// a breadth first search for paths against a stellar-core's database
+package simplepath
