+++ src/github.com/stellar/horizon/simplepath/finder.go
@@ -1,29 +1,28 @@
-package db
+package simplepath
 
 import (
-	"bytes"
-	"fmt"
 	"github.com/go-errors/errors"
 	"github.com/stellar/go-stellar-base/xdr"
+	"github.com/stellar/horizon/db"
 	"github.com/stellar/horizon/log"
 	"github.com/stellar/horizon/paths"
 	"golang.org/x/net/context"
 )
 
-// SimplePathFinder implements the paths.Finder interface and searchs for
+// Finder implements the paths.Finder interface and searchs for
 // payment paths using a simple breadth first search of the offers table of a stellar-core.
 //
 // This implementation is not meant to be fast or to provide the lowest costs paths, but
 // rather is meant to be a simple implementation that gives usable paths.
-type SimplePathFinder struct {
-	SqlQuery
+type Finder struct {
+	db.SqlQuery
 	Ctx context.Context
 }
 
-// ensure SimplePathFinder is paths.Finder compliant
-var _ paths.Finder = &SimplePathFinder{}
+// ensure the struct is paths.Finder compliant
+var _ paths.Finder = &Finder{}
 
-func (f *SimplePathFinder) Find(q paths.Query) (result []paths.Path, err error) {
+func (f *Finder) Find(q paths.Query) (result []paths.Path, err error) {
 	log.WithField(f.Ctx, "source_assets", q.SourceAssets).
 		WithField("destination_asset", q.DestinationAsset).
 		WithField("destination_amount", q.DestinationAmount).
@@ -71,8 +70,8 @@ func (f *SimplePathFinder) Find(q paths.Query) (result []paths.Path, err error) 
 		}
 
 		var connected []xdr.Asset
-		q := AssetsWithDepthQuery{f.SqlQuery, cur.Asset, int64(minDepth)}
-		err = Select(f.Ctx, q, &connected)
+		q := db.AssetsWithDepthQuery{f.SqlQuery, cur.Asset, int64(minDepth)}
+		err = db.Select(f.Ctx, q, &connected)
 
 		if err != nil {
 			return
@@ -84,84 +83,5 @@ func (f *SimplePathFinder) Find(q paths.Query) (result []paths.Path, err error) 
 	}
 
 	log.WithField(f.Ctx, "found", len(result)).Info("Finished pathfind")
-	return
-}
-
-type pathNode struct {
-	Asset xdr.Asset
-
-	Tail *pathNode
-}
-
-func (p *pathNode) String() string {
-	if p == nil {
-		return ""
-	}
-
-	var out bytes.Buffer
-	fmt.Fprintf(&out, "%v", p.Asset)
-
-	cur := p.Tail
-
-	for cur != nil {
-		fmt.Fprintf(&out, " -> %v", cur.Asset)
-		cur = cur.Tail
-	}
-
-	return out.String()
-}
-
-func (p *pathNode) Source() xdr.Asset {
-	cur := p
-	for cur.Tail != nil {
-		cur = cur.Tail
-	}
-	return cur.Asset
-}
-
-func (p *pathNode) Destination() xdr.Asset {
-	// the destination for path is the head of the linked list
-	return p.Asset
-}
-
-func (p *pathNode) Path() []xdr.Asset {
-	path := p.Flatten()
-
-	if len(path) < 2 {
-		return nil
-	}
-
-	// return the flattened slice without the first and last elements
-	// which are the source and the destination assets
-	return path[1 : len(path)-1]
-}
-
-func (p *pathNode) Cost(amount xdr.Int64) xdr.Int64 {
-	return amount
-}
-
-func (p *pathNode) Depth() int {
-	depth := 0
-	cur := p
-	for {
-		if cur == nil {
-			return depth
-		}
-		cur = cur.Tail
-		depth++
-	}
-}
-
-func (p *pathNode) Flatten() (result []xdr.Asset) {
-	cur := p
-
-	for {
-		if cur == nil {
-			return
-		}
-		result = append(result, cur.Asset)
-		cur = cur.Tail
-	}
-
 	return
 }
